Use project.json instead of packages.config

+++ src/IronPigeon/CryptoProviderExtensions.cs
@@ -228,8 +228,8 @@
 		/// A set of encryption variables.
 		/// </returns>
 		private static SymmetricEncryptionVariables NewSymmetricEncryptionVariables(CryptoSettings cryptoProvider) {
-			byte[] key = WinRTCrypto.CryptographicBuffer.GenerateRandom((uint)cryptoProvider.SymmetricKeySize / 8);
-			byte[] iv = WinRTCrypto.CryptographicBuffer.GenerateRandom((uint)CryptoSettings.SymmetricAlgorithm.BlockLength);
+			byte[] key = WinRTCrypto.CryptographicBuffer.GenerateRandom(cryptoProvider.SymmetricKeySize / 8);
+			byte[] iv = WinRTCrypto.CryptographicBuffer.GenerateRandom(CryptoSettings.SymmetricAlgorithm.BlockLength);
 			return new SymmetricEncryptionVariables(key, iv);
 		}
